scan_sources/requirements.txt
crewai>=0.30.11
crewai_tools>=0.2.6
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
# Add other specific tool dependencies here as needed, e.g.:
# serperapi
# beautifulsoup4
# requests
# PyPDF2
